Added Country Languages api.

+++ src/main/java/com/project/gridbackend/repository/CountryRepository.java
@@ -3,5 +3,5 @@ package com.project.gridbackend.repository;
 import com.project.gridbackend.model.Country;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface CountryRepository extends PagingAndSortingRepository<Country, Long> {
+public interface CountryRepository extends PagingAndSortingRepository<Country, Integer> {
 }
